refactor(peer-cleanup): trait-based driver + RFCOMM tracker impl
Introduces the central peer-cleanup driver that the next commits
plug additional trackers into:

  * `pim-daemon/src/peer_cleanup.rs` — the `PeerCleanupTracker`
    trait and `spawn` driver. The driver clamps lifetime and
    interval to `MIN_LIFETIME_S` / `MIN_INTERVAL_S` (1 h / 60 s),
    skips altogether when the policy's `enabled = false`, logs
    every tick failure at WARN, and is cancellable via the same
    token the rest of the daemon uses for shutdown.

  * `pim-daemon/src/peer_cleanup/rfcomm.rs` — `RfcommTracker`
    impl. Same logic as the previous standalone
    `rfcomm_cleanup` module: list lifecycle rows, drop dangling
    ones whose peer is no longer paired in BlueZ, unpair via
    `bluetoothctl remove` after the lifetime threshold + a
    `Connected: no` confirmation. Exports moved to
    `crate::app::peer_cleanup::{spawn, MIN_LIFETIME_S}`.

`pim-daemon/src/rfcomm_cleanup.rs` is removed; its tests live
inside the new `rfcomm.rs` and continue to cover the rejoin /
past-lifetime / currently-connected scenarios. `pim-core` now
re-exports `PeerCleanupConfig` at the crate root so daemon-side
callers don't have to dig into `pim_core::config`.

Spawn site in `app.rs` constructs an `Arc<dyn
PeerCleanupTracker>` from the RFCOMM impl and passes it to the
driver. Behaviour unchanged.
